+++ packages/plugin-commons/web/src/Base/createBasePlugin.ts
@@ -7,7 +7,6 @@ import { generateInsertPluginButtonProps } from '../Utils/generateInsertPluginBu
 import {
   deleteBlock,
   setBlockNewEntityData,
-  setEntityData,
   getToolbarTheme,
   TOOLBARS,
 } from 'wix-rich-content-editor-common';
@@ -50,12 +49,9 @@ const getData = (
 const setData = (
   contentBlock: ContentBlock,
   { getEditorState, setEditorState }: EditorStateFuncs,
-  type: string,
-  commonPubsub
+  type: string
 ) => newData => {
-  const editorState = commonPubsub.get('undoExperiment')?.()
-    ? setBlockNewEntityData(getEditorState(), contentBlock.getKey(), newData, type)
-    : setEntityData(getEditorState(), contentBlock.getEntityAt(0), newData);
+  const editorState = setBlockNewEntityData(getEditorState(), contentBlock.getKey(), newData, type);
   setEditorState(editorState);
 };
 
@@ -313,12 +309,7 @@ const createBasePlugin = (
                   { getEditorState },
                   type === UNSUPPORTED_BLOCKS_TYPE
                 ),
-                setData: setData(
-                  contentBlock,
-                  { getEditorState, setEditorState },
-                  config.type,
-                  commonPubsub
-                ),
+                setData: setData(contentBlock, { getEditorState, setEditorState }, config.type),
                 deleteBlock: deleteEntity(contentBlock, { getEditorState, setEditorState }),
                 type,
               },